+++ lib/app/my_app.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:novel_v3/app/ui/get_start/getstart_screen.dart';
 import 'package:novel_v3/more_libs/setting_v2.0.0/setting.dart';
 import 'package:t_server/core/t_server.dart';
+import 'package:than_pkg/than_pkg.dart';
 
 import 'ui/home_screen.dart';
 
@@ -21,13 +23,16 @@ class _MyAppState extends State<MyApp> {
 
   @override
   Widget build(BuildContext context) {
+    final readGetstart = TRecentDB.getInstance.getBool('read-getstart');
     return ThemeSwitcher(
       builder: (config) {
         return MaterialApp(
           debugShowCheckedModeBanner: false,
           darkTheme: ThemeData.dark(useMaterial3: true),
           themeMode: config.isDarkMode ? ThemeMode.dark : ThemeMode.light,
-          home: const HomeScreen(),
+          home: readGetstart
+              ? const HomeScreen()
+              : GetstartScreen(child: HomeScreen()),
         );
       },
     );