+++ lib/app/ui/main_ui/screens/home/style_pages/home_grid_style.dart
@@ -8,6 +8,7 @@ import 'package:novel_v3/app/ui/main_ui/screens/forms/edit_novel_form.dart';
 import 'package:novel_v3/app/ui/routes_helper.dart';
 import 'package:provider/provider.dart';
 import 'package:t_widgets/t_widgets_dev.dart';
+import 'package:than_pkg/t_database/t_recent_db.dart';
 
 class HomeGridStyle extends StatefulWidget {
   const HomeGridStyle({super.key});
@@ -35,6 +36,17 @@ class _HomeGridStyleState extends State<HomeGridStyle> {
   void initState() {
     filterName = filter.first;
     super.initState();
+    init();
+  }
+
+  void init() async {
+    final name = TRecentDB.getInstance.getString(
+      'home-grid-style-filter-name',
+      def: filter.first,
+    );
+    setState(() {
+      filterName = name;
+    });
   }
 
   @override
@@ -43,7 +55,12 @@ class _HomeGridStyleState extends State<HomeGridStyle> {
     if (list.isEmpty) {
       return _getEmptyList();
     }
-    return CustomScrollView(slivers: [_getHeader(), _getList(list)]);
+    return RefreshIndicator.adaptive(
+      onRefresh: () async {
+        context.read<NovelProvider>().initList(isCached: false);
+      },
+      child: CustomScrollView(slivers: [_getHeader(), _getList(list)]),
+    );
   }
 
   Widget _getEmptyList() {
@@ -80,6 +97,10 @@ class _HomeGridStyleState extends State<HomeGridStyle> {
                 setState(() {
                   filterName = name;
                 });
+                TRecentDB.getInstance.putString(
+                  'home-grid-style-filter-name',
+                  name,
+                );
               },
             );
           }),